requirements.txt
dateutil
certifi==2019.9.11
Click==7.0
Cython==0.29.14
Django==2.2.7
Flask==1.1.1
gunicorn==20.0.0
image==1.5.27
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
olefile==0.46
Pillow==6.2.1
pytesseract==0.3.0
pytz==2019.3
sqlparse==0.3.0
tesseract==0.1.3
tesseract-ocr==0.0.1
Werkzeug==0.16.0
